+++ lib/src/fetch.dart
@@ -4,15 +4,25 @@ import 'package:gotrue/src/fetch_options.dart';
 import 'package:gotrue/src/gotrue_error.dart';
 import 'package:gotrue/src/gotrue_response.dart';
 import 'package:http/http.dart' as http;
-
-final Fetch fetch = Fetch();
+import 'package:http/http.dart';
+import 'package:universal_io/io.dart';
 
 class Fetch {
+  final Client? httpClient;
+
+  Fetch([this.httpClient]);
+
   bool isSuccessStatusCode(int code) {
     return code >= 200 && code <= 299;
   }
 
-  GotrueError handleError(dynamic error) {
+  GotrueResponse handleError(dynamic error) {
+    int? statusCode;
+    GotrueError errorRes;
+    if (error is http.Response) {
+      statusCode = error.statusCode;
+    }
+
     if (error is http.Response) {
       try {
         final parsedJson = json.decode(error.body) as Map<String, dynamic>;
@@ -21,35 +31,47 @@ class Fetch {
             parsedJson['error_description'] ??
             parsedJson['error'] ??
             json.encode(parsedJson);
-        return GotrueError(message as String);
+        errorRes = GotrueError(message as String);
       } on FormatException catch (_) {
-        return GotrueError(error.body);
+        errorRes = GotrueError(error.body);
       }
+    } else if (error is SocketException) {
+      errorRes = GotrueError(
+        error.toString(),
+        statusCode: error.runtimeType.toString(),
+      );
     } else {
-      return GotrueError(error.toString());
+      errorRes = GotrueError(error.toString());
     }
+
+    return GotrueResponse(error: errorRes, statusCode: statusCode);
   }
 
   Future<GotrueResponse> get(String url, {FetchOptions? options}) async {
-    final client = http.Client();
     try {
       final headers = options?.headers ?? {};
-      final http.Response response =
-          await client.get(Uri.parse(url), headers: headers);
+      final http.Response response = await (httpClient?.get ?? http.get)(
+        Uri.parse(url),
+        headers: headers,
+      );
       if (isSuccessStatusCode(response.statusCode)) {
         if (options?.noResolveJson == true) {
-          return GotrueResponse(rawData: response.body);
+          return GotrueResponse(
+            rawData: response.body,
+            statusCode: response.statusCode,
+          );
         } else {
           final jsonBody = json.decode(response.body);
-          return GotrueResponse(rawData: jsonBody);
+          return GotrueResponse(
+            rawData: jsonBody,
+            statusCode: response.statusCode,
+          );
         }
       } else {
         throw response;
       }
     } catch (e) {
-      return GotrueResponse(error: handleError(e));
-    } finally {
-      client.close();
+      return handleError(e);
     }
   }
 
@@ -58,26 +80,34 @@ class Fetch {
     dynamic body, {
     FetchOptions? options,
   }) async {
-    final client = http.Client();
     try {
       final bodyStr = json.encode(body ?? {});
       final headers = options?.headers ?? {};
-      final http.Response response =
-          await client.post(Uri.parse(url), headers: headers, body: bodyStr);
+      final http.Response response = await (httpClient?.post ?? http.post)(
+        Uri.parse(url),
+        headers: headers,
+        body: bodyStr,
+      );
       if (isSuccessStatusCode(response.statusCode)) {
         if (options?.noResolveJson == true) {
-          return GotrueResponse(rawData: response.body);
+          return GotrueResponse(
+            rawData: response.body,
+            statusCode: response.statusCode,
+          );
         } else {
           final jsonBody = json.decode(response.body);
-          return GotrueResponse(rawData: jsonBody);
+          return GotrueResponse(
+            rawData: jsonBody,
+            statusCode: response.statusCode,
+          );
         }
       } else {
         throw response;
       }
+    } on SocketException catch (e) {
+      return handleError(e);
     } catch (e) {
-      return GotrueResponse(error: handleError(e));
-    } finally {
-      client.close();
+      return handleError(e);
     }
   }
 
@@ -86,26 +116,32 @@ class Fetch {
     dynamic body, {
     FetchOptions? options,
   }) async {
-    final client = http.Client();
     try {
       final bodyStr = json.encode(body ?? {});
       final headers = options?.headers ?? {};
-      final http.Response response =
-          await client.put(Uri.parse(url), headers: headers, body: bodyStr);
+      final http.Response response = await (httpClient?.put ?? http.put)(
+        Uri.parse(url),
+        headers: headers,
+        body: bodyStr,
+      );
       if (isSuccessStatusCode(response.statusCode)) {
         if (options?.noResolveJson == true) {
-          return GotrueResponse(rawData: response.body);
+          return GotrueResponse(
+            rawData: response.body,
+            statusCode: response.statusCode,
+          );
         } else {
           final jsonBody = json.decode(response.body);
-          return GotrueResponse(rawData: jsonBody);
+          return GotrueResponse(
+            rawData: jsonBody,
+            statusCode: response.statusCode,
+          );
         }
       } else {
         throw response;
       }
     } catch (e) {
-      return GotrueResponse(error: handleError(e));
-    } finally {
-      client.close();
+      return handleError(e);
     }
   }
 }