+++ lib/src/gotrue_error.dart
@@ -1,10 +1,14 @@
 class GotrueError {
   String message;
+  String? statusCode;
 
-  GotrueError(this.message);
+  GotrueError(
+    this.message, {
+    this.statusCode,
+  });
 
   @override
   String toString() {
-    return 'GotrueError(message: $message)';
+    return 'GotrueError(message: $message, statusCode: $statusCode)';
   }
 }